pyproject.toml
[project]
name = "ai-knowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi==0.115.0",
    "grandalf>=0.8",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "mcp>=1.12.4",
    "openai>=2.15.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]==0.30.6",
]
